feat: add a generic way to insert a struct into the database
database/create - a generic function that is capable to insert anything into the database. Needs the order of colomn names to match the values
handlers/posts/create.go - a handler that serves a html page with a form to create a post, and also handles the posted form on submit. It utilizes the generic function above to perform its work
utils/random_string.go - generates a random string. We need this functionality in order to save images using predictable names
media/ - will be created if it does not exist st the root of the project. It is where images are to be stored
added a "created_at" field on posts and comments table. Although they get initialized by the database by default